Всё рабоатет в обоих дизайнах
