+++ src/pages/ArticlesPage/components/ArticlesPage.tsx
@@ -1,18 +1,11 @@
 import { memo } from 'react';
-import { useAppDispatch } from '@/shared/lib/hooks/useAppDispatch/useAppDispatch';
 import { useSelector } from 'react-redux';
 import { PageMainContent } from '@/widgets/PageMainContent';
-import {
-	ArticlesInfiniteList,
-	getArticlesInfiniteListError,
-	getArticlesInfiniteListIsLoading,
-} from '@/widgets/articlesInfiniteList';
+import { ArticlesInfiniteList, getArticlesInfiniteListError } from '@/widgets/articlesInfiniteList';
 import { ArticlesPageGreeting } from '@/features/articlesPageGreeting';
 
 export const ArticlesPage = memo(() => {
-	const dispatch = useAppDispatch();
 	const error = useSelector(getArticlesInfiniteListError);
-	const isLoading = useSelector(getArticlesInfiniteListIsLoading);
 
 	return (
 		<PageMainContent data-testid='ArticlesPage'>
