+++ src/widgets/articlesInfiniteList/index.ts
@@ -1,7 +1,4 @@
 export { ArticlesInfiniteList } from './components/ArticlesInfiniteList/ArticlesInfiniteList';
 export type { ArticlesInfiniteListSchema } from './model/types/articlesInfiniteListSchema';
-export {
-	getArticlesInfiniteListError,
-	getArticlesInfiniteListIsLoading,
-} from './model/selectors/articlesInfiniteListSelectors';
+export { getArticlesInfiniteListError } from './model/selectors/articlesInfiniteListSelectors';
 export { fetchNextArticlesPage } from './model/services/fetchNextArticlesPage/fetchNextArticlesPage';
